+++ libc/string.cpp
@@ -25,7 +25,7 @@
 
 extern "C" void *memset(void *ptr, int value, size_t num)
 {
-    for (int i = 0; i < num; i++)
+    for (size_t i = 0; i < num; i++)
     {
         ((char *)ptr)[i] = (char)(value);
     }
